requirements.txt
requests #to download images
spotipy #to access spotify
pillow
dotenv
